requirements.txt
numpy
pandas
matplotlib
tqdm
kagglehub
paho-mqtt
loguru

Django
djangorestframework
django-cors-headers
paho-mqtt
python-dotenv
drf-spectacular
